split `views/GetStartedPage/CreateWalletPage/Create.module.css + improve declarativity
