+++ app/components/views/GetStartedPage/CreateWalletPage/ConfirmSeed/Form/index.js
@@ -1,0 +1,1 @@
+export { default } from "./Form";
