+++ app/components/views/GetStartedPage/CreateWalletPage/CopySeed/Form/index.js
@@ -1,0 +1,1 @@
+export { default } from "./Form";
