--- app/components/views/GetStartedPage/CreateWalletPage/CreateWallet.module.css
@@ -1,321 +1,0 @@
-.goBackScreenButton {
-  height: 12px;
-  width: 12px;
-  background-repeat: no-repeat;
-  background-size: 12px;
-  background-image: var(--x-grey);
-  opacity: 1;
-  cursor: pointer;
-}
-
-.goBackScreenButton:hover {
-  opacity: 0.7;
-}
-
-.createWalletWarning {
-  column-count: 2;
-  max-width: 48em;
-  font-size: 13px;
-  line-height: 16px;
-  margin-bottom: 1em;
-  color: var(--new_wallet-label);
-}
-
-.createWalletWarning p {
-  margin-top: 0;
-}
-.createWalletWarning em {
-  color: var(--error-text-color);
-}
-.createWalletWarning strong {
-  color: var(--title-text-and-button-background);
-}
-
-.seedError {
-  margin: -10px 0 20px;
-  color: var(--error-text-color);
-}
-
-.seedWarning {
-  margin: -10px 0 20px;
-  color: var(--error-text-color);
-}
-
-.seedWord {
-  user-select: none;
-  height: 23px;
-  border-radius: 3px;
-  color: var(--info-modal-button-text);
-  margin: 5px;
-  text-align: center;
-  display: flex;
-  align-items: center;
-}
-
-.seedWord.filled {
-  background-color: var(--info-modal-button-bg);
-}
-
-.seedWord.empty,
-.seedWord.restore {
-  border: 1px solid var(--seedword-select-border-default-border-color);
-  background-color: initial !important;
-}
-
-.seedWord.empty:hover,
-.seedWord.restore:hover {
-  border: 1px solid var(--seedword-select-border-default-hover-border-color);
-}
-
-.seedWord.empty:hover .number,
-.seedWord.restore:hover .number {
-  border-right: 1px solid
-    var(--seedword-select-border-default-hover-border-color);
-  color: var(--seedword-select-border-default-hover-text-color);
-}
-
-.seedWord.empty .number,
-.seedWord.restore .number {
-  border-right: 1px solid var(--seedword-select-border-default-border-color);
-  color: var(--seedword-select-border-default-text-color);
-}
-
-.seedWord.populated {
-  border: 1px solid var(--seedword-select-border-default-hover-text-color);
-  background-color: initial !important;
-}
-
-.seedWord.populated .number {
-  border-right: 1px solid var(--seedword-select-border-default-hover-text-color);
-  color: var(--seedword-select-border-default-hover-text-color);
-}
-
-.seedWord.noMatch,
-.seedWord.error {
-  border: 1px solid var(--error-red);
-}
-
-.seedWord.noMatch .number,
-.seedWord.error .number {
-  border-right: 1px solid var(--error-red);
-  color: var(--error-red);
-}
-
-.seedWord.match {
-  border: 1px solid var(--accent-color);
-}
-
-.seedWord.match .number {
-  border-right: 1px solid var(--accent-color);
-  color: var(--accent-color);
-}
-
-.copy {
-  grid-column: 4 / 6;
-  padding-top: 2px;
-  padding-left: 43px;
-  height: 24px;
-  color: var(--input-color);
-  border: 1px solid var(--input-color);
-  border-radius: 3px;
-  margin: 4px;
-  font-size: 13px;
-  background-image: var(--copy-to-clipboard-icon);
-  background-position: 26px 3px;
-  background-size: 16px auto;
-  background-repeat: no-repeat;
-  cursor: pointer;
-  user-select: none;
-}
-
-.copy:hover {
-  opacity: 0.8;
-}
-
-.walletKeyBlueButton {
-  float: left;
-}
-
-.seed {
-  margin-bottom: 30px;
-}
-
-.confirmSeedLabel {
-  margin-top: 8px;
-  margin-right: 10px;
-  font-size: 13px;
-  line-height: 17px;
-  width: 150px;
-  margin-left: -160px;
-  color: var(--wallet-tutorial-text);
-  justify-content: flex-end;
-  text-align: right;
-}
-
-.createWalletButtonContainer {
-  height: 44px;
-  width: 293px;
-  display: flex;
-  flex-direction: row-reverse;
-}
-
-.getstartedPageBody .textToggle {
-  height: 29px;
-}
-
-/* TODO: shorter classes here please when done migrating!!!! */
-
-.textToggleButtonLeft {
-  color: var(--text-toggle-button-inactive-text);
-  background-color: var(--text-toggle-button-bg);
-  border: 1px solid var(--text-toggle-button-border);
-  border-radius: 3px 0px 0px 3px;
-  min-width: 0px;
-  padding: 6px 20px 6px 20px;
-}
-
-.textToggleButtonRight {
-  color: var(--text-toggle-button-inactive-text);
-  background-color: var(--text-toggle-button-bg);
-  border-radius: 0px 3px 3px 0px;
-  min-width: 0px;
-  padding: 6px 20px 6px 20px;
-}
-
-.textToggleButtonLeft.textToggleButtonActive {
-  color: var(--input-color-default);
-  background-color: var(--text-toggle-button-active-bg);
-  font-weight: 400;
-}
-
-.textToggleButtonRight.textToggleButtonActive {
-  color: var(--input-color-default);
-  background-color: var(--text-toggle-button-active-bg);
-  font-weight: 400;
-}
-
-.goBackButton {
-  color: var(--home-content-link);
-}
-.goBackButton:hover {
-  color: var(--title-text-and-button-background-hovered);
-}
-
-.seedArea {
-  margin: 0px 5px 5px -5px;
-  width: 560px;
-  display: grid;
-  grid-template-columns: repeat(5, 1fr);
-}
-
-.number {
-  width: 20px;
-  border-right: 1px solid var(--seedword-number-color);
-  color: var(--seedword-number-color);
-  font-size: 10px;
-  height: 23px;
-  line-height: 23px;
-}
-
-.word {
-  text-align: center;
-  padding-left: 5px;
-  width: 85px;
-  font-size: 13px;
-  line-height: 23px;
-}
-
-.seedArea.hex {
-  width: 90%;
-  display: block;
-}
-
-.buttonArea {
-  margin-left: 22px;
-  margin-top: 189px;
-  width: 380px;
-}
-
-.button {
-  background-color: var(--background-back-color);
-  border: 1px solid var(--background-back-color);
-  border-radius: 5px;
-  height: 161px;
-  width: 168px;
-  margin-right: 20px;
-  float: left;
-  background-position: 53px 32px;
-  background-repeat: no-repeat;
-  background-size: 62px 62px;
-}
-
-.button.new {
-  background-image: var(--createnewwallet);
-}
-
-.button.restore {
-  background-image: var(--restore-wallet-icon);
-}
-
-.button:hover,
-.button:focus {
-  border-color: var(--input-color);
-}
-
-.buttonLabel {
-  margin-top: 121px;
-  text-align: center;
-  font-size: 15px;
-  color: var(--title-text-and-button-background);
-  line-height: 19px;
-}
-
-.walletSwitch {
-  padding-right: 5px;
-}
-
-.toolbar {
-  display: flex;
-  flex-direction: row-reverse;
-  width: 596px;
-  margin-top: 20px;
-  align-items: center;
-}
-
-@media screen and (max-width: 1179px) {
-  .seed {
-    margin-left: 0;
-  }
-
-  .passphrase {
-    margin-left: 0;
-  }
-}
-
-@media screen and (max-width: 768px) {
-  .seed {
-    flex-direction: column;
-  }
-
-  .toolbar {
-    width: 380px !important;
-    margin-left: -20px;
-  }
-
-  .copy {
-    grid-column: 1 / 4;
-    background-position: 80px;
-    text-align: center;
-  }
-
-  .seedArea {
-    grid-template-columns: repeat(3, 1fr);
-    width: 353px;
-  }
-
-  .createWalletButtonContainer {
-    display: flex;
-    justify-content: flex-end;
-    width: 365px;
-  }
-}
