+++ app/components/views/GetStartedPage/CreateWalletPage/ExistingSeed/Form/Form.module.css
@@ -1,0 +1,16 @@
+.hex {
+  width: 90%;
+  display: block;
+}
+
+.confirmSeedLabel {
+  margin-top: 8px;
+  margin-right: 10px;
+  font-size: 13px;
+  line-height: 17px;
+  width: 150px;
+  margin-left: -160px;
+  color: var(--wallet-tutorial-text);
+  justify-content: flex-end;
+  text-align: right;
+}
