+++ app/components/views/GetStartedPage/CreateWalletPage/ExistingSeed/Form/index.js
@@ -1,0 +1,1 @@
+export { default } from "./Form";
