+++ app/components/views/GetStartedPage/CreateWalletPage/helpers/ButtonsBar/index.js
@@ -1,0 +1,1 @@
+export { default } from "./ButtonsBar";
