+++ app/components/views/GetStartedPage/CreateWalletPage/helpers/SeedArea/SeedArea.jsx
@@ -1,0 +1,7 @@
+import { classNames } from "pi-ui";
+import styles from "./SeedArea.module.css";
+
+const SeedArea = ({ className, children }) => (
+  <div className={classNames(styles.seedArea, className)}>{children}</div>
+);
+export default SeedArea;
