+++ app/components/views/GetStartedPage/CreateWalletPage/helpers/SeedArea/index.js
@@ -1,0 +1,1 @@
+export { default } from "./SeedArea";
