+++ app/components/views/GetStartedPage/CreateWalletPage/helpers/SeedError/SeedError.jsx
@@ -1,0 +1,6 @@
+import styles from "./SeedError.module.css";
+
+const SeedError = ({ children }) => (
+  <div className={styles.seedError}>{children}</div>
+);
+export default SeedError;
