+++ app/components/views/GetStartedPage/CreateWalletPage/helpers/SeedError/SeedError.module.css
@@ -1,0 +1,4 @@
+.seedError {
+  margin: -10px 0 20px;
+  color: var(--error-text-color);
+}
