+++ app/components/views/GetStartedPage/CreateWalletPage/helpers/SeedError/index.js
@@ -1,0 +1,1 @@
+export { default } from "./SeedError";
