+++ app/components/views/GetStartedPage/CreateWalletPage/helpers/SeedWord/index.js
@@ -1,0 +1,1 @@
+export { default } from "./SeedWord";
