+++ app/components/views/GetStartedPage/CreateWalletPage/helpers/SeedWordEntry/SeedWordSelect.css
