+++ app/components/views/GetStartedPage/CreateWalletPage/helpers/SeedWordEntry/index.js
@@ -1,0 +1,1 @@
+export { default } from "./SeedWordEntry";
