+++ app/components/views/GetStartedPage/CreateWalletPage/helpers/index.js
@@ -1,2 +1,7 @@
 export { default as Container } from "./Container";
 export { default as TitleWrapper } from "./TitleWrapper";
+export { default as Section } from "./Section";
+export { default as SeedArea } from "./SeedArea";
+export { default as SeedWord } from "./SeedWord";
+export { default as SeedError } from "./SeedError";
+export { default as ButtonsBar } from "./ButtonsBar";
