+++ app/components/views/GetStartedPage/Logs/Logs.jsx
@@ -2,12 +2,13 @@ import { Tooltip } from "pi-ui";
 import { LogsTab } from "views/SettingsPage/LogsTab/LogsTab";
 import { GoBackMsg } from "../messages";
 import styles from "../GetStarted.module.css";
+import { BackButton } from "../helpers";
 
 export default ({ onSendBack }) => (
   <>
     <div className={styles.goBackScreenButtonArea}>
       <Tooltip content={<GoBackMsg />}>
-        <div className={styles.goBackScreenButton} onClick={onSendBack} />
+        <BackButton onClick={onSendBack} />
       </Tooltip>
     </div>
     <div className={styles.logContainer}>
