+++ app/components/views/GetStartedPage/helpers/BackButton/BackButton.jsx
@@ -1,0 +1,7 @@
+import styles from "./BackButton.module.css";
+
+const BackButton = ({ onClick }) => (
+  <div onClick={onClick} className={styles.backButton}></div>
+);
+
+export default BackButton;
