+++ app/components/views/GetStartedPage/helpers/BackButton/BackButton.module.css
@@ -1,0 +1,13 @@
+.backButton {
+  height: 12px;
+  width: 12px;
+  background-repeat: no-repeat;
+  background-size: 12px;
+  background-image: var(--x-grey);
+  opacity: 1;
+  cursor: pointer;
+}
+
+.backButton:hover {
+  opacity: 0.7;
+}
