+++ app/components/views/GetStartedPage/helpers/BackButton/index.js
@@ -1,0 +1,1 @@
+export { default } from "./BackButton";
