+++ app/components/views/GetStartedPage/helpers/index.js
@@ -3,3 +3,4 @@ export { default as Input } from "./Input";
 export { default as FormContainer } from "./FormContainer";
 export { default as Row } from "./Row";
 export { default as ButtonsBar } from "./ButtonsBar";
+export { default as BackButton } from "./BackButton";
